move fade in/out functions to util
